Rename and move storage-lib
