+++ packages/client/src/modules/app/epics.ts
@@ -10,7 +10,6 @@ import {
 import { graphqlSelector } from "graphql/queries";
 import Analytics from "lib/analytics-lib";
 import ENV from "lib/client-env";
-import StorageModule from "lib/storage-lib";
 import { EpicSignature } from "modules/root";
 import { i18nSelector } from "modules/settings/selectors";
 import { combineEpics } from "redux-observable";
@@ -31,6 +30,7 @@ import {
   isChartTabValidForNetwork,
   onChartTab,
 } from "tools/client-utils";
+import StorageModule from "tools/storage-utils";
 import {
   validateEmailAddress,
   validateNetworkAddress,
